Data transfer backend logic implementation.

+++ webstack/apps/homebase/src/api/collections/users.ts
@@ -9,7 +9,7 @@
 import { UserSchema } from '@sage3/shared/types';
 import { SAGE3Collection, sageRouter } from '@sage3/backend';
 
-import { SAGEBase } from '@sage3/sagebase';
+import { SAGEBase, SBAuthSchema } from '@sage3/sagebase';
 
 import { config } from '../../config';
 import { RoomsCollection } from './rooms';
@@ -17,6 +17,7 @@ import { BoardsCollection } from './boards';
 import { AppsCollection } from './apps';
 import { AssetsCollection } from './assets';
 import { PluginsCollection } from './plugins';
+import { RoomMembersCollection } from './roommembers';
 
 type UserStats = {
   userId: string;
@@ -80,9 +81,10 @@ class SAGE3UsersCollection extends SAGE3Collection<UserSchema> {
     });
 
     router.post('/accountDeletion', async ({ body, user }, res) => {
-      const userId = (user as any).id;
-      console.log(userId);
-      if (!userId) {
+      const userRequestingDeletion = user as SBAuthSchema;
+
+      // Is the user logged in?
+      if (!userRequestingDeletion.id) {
         res.status(401).send({ success: false, message: 'Unauthorized' });
         return;
       }
@@ -90,36 +92,43 @@ class SAGE3UsersCollection extends SAGE3Collection<UserSchema> {
       // Does the Body contain the id to delete?
       const userIdToDelete = body.id;
       if (!userIdToDelete) {
-        res.status(400).send({ success: false, message: 'No ID provided' });
+        res.status(400).send({ success: false, message: 'No User ID provided' });
         return;
       }
 
-      // Delete all data?
-      const deleteAllData = body.deleteAllData ? body.deleteAllData : false;
-
-      // Only admins and the user themselves can delete their account
+      // Is the user making the rqeuest the same as the user to delete?
       let isSameUser = false;
-      if (userId === userIdToDelete) {
+      if (userRequestingDeletion.id === userIdToDelete) {
         isSameUser = true;
       }
 
       // Get the user from the database
-      const userDoc = await this.get(userIdToDelete);
-      // Get the user email
-      const userEmail = userDoc?.data.email;
+      const userToDelete = await this.get(userIdToDelete);
 
-      // If no user email was found, return an error
-      if (!userEmail) {
+      // Does the user exist?
+      if (!userToDelete) {
+        res.status(404).send({ success: false, message: 'User not found' });
+        return;
+      }
+
+      // Does ther user have an email? If not, they are a guest and cannot be deleted
+      if (!userToDelete.data.email) {
         res.status(404).send({ success: false, message: 'User is not registered and may be a guest.' });
         return;
       }
 
-      // Check if the user who is requesting the deletion is an admin
-      // Get Admins from config file
+      // Check if the user to be deleted is an admin
       const adminEmails = config.auth.admins || [];
+      const userEmail = userToDelete.data.email;
+      if (adminEmails.includes(userEmail)) {
+        res.status(401).send({ success: false, message: 'Unauthorized. Admins can not be deleted.' });
+        return;
+      }
+
+      // Check if the user who is requesting the deletion is an admin
       let isAdmin = false;
-      const userPerformingDeletion = await this.get(userId);
-      const userEmailPerformingDeletion = userPerformingDeletion?.data.email;
+      const userPerformingDeletion = await this.get(userRequestingDeletion.id);
+      const userEmailPerformingDeletion = userPerformingDeletion ? userPerformingDeletion.data.email : null;
       if (userEmailPerformingDeletion) {
         isAdmin = adminEmails.includes(userEmailPerformingDeletion);
       }
@@ -130,6 +139,24 @@ class SAGE3UsersCollection extends SAGE3Collection<UserSchema> {
         return;
       }
 
+      // Check if the user wants to delete all data
+      const deleteAllData = body.deleteAllData ? body.deleteAllData : false;
+      // If the user doesnt want to delete their data, ensure there is an admin account to transfer the data to
+      const adminEmail = adminEmails[0];
+      if (!deleteAllData && !adminEmail) {
+        res.status(400).send({ success: false, message: 'No admin account to transfer data to. Please contact the server administrator.' });
+        return;
+      }
+      const adminAccounts = await this.query('email', adminEmail);
+      const adminAccount = adminAccounts ? adminAccounts[0] : null;
+      if (!deleteAllData && !adminAccount) {
+        res
+          .status(400)
+          .send({ success: false, message: 'Admin account not found to transfer data to. Please contact the server administrator.' });
+        return;
+      }
+
+      // All checks passed. Start the deletion process.
       this.userDeleteLog('User Deletion Request Started', userEmail);
       this.userDeleteLog(`Delete All Data: ${deleteAllData}`, userEmail);
 
@@ -141,10 +168,40 @@ class SAGE3UsersCollection extends SAGE3Collection<UserSchema> {
       const authCollectionRemoval = authCollDelete ? true : false;
       this.userDeleteLog(`Auth Collection Delete ${authCollectionRemoval ? 'Success' : 'Failed'}`, userEmail);
 
-      if (!deleteAllData) {
-        this.userDeleteLog('User data will not be deleted and will be transfered to the server admin.', userEmail);
+      // If the data is not to be deleted, transfer ownership of rooms, boards, apps, assets, and plugins to the server admins
+      if (!deleteAllData && adminAccount) {
+        const adminId = adminAccount._id;
 
-        // TODO Logic to transfer ownership of rooms, boards, apps, assets, and plugins to the server admins
+        this.userDeleteLog(`User data will be transferred to admin account with email: ${adminAccount.data.email}`, userEmail);
+
+        // Transfer Rooms
+        const roomsTransfer = await RoomsCollection.transferUsersRooms(userIdToDelete, adminId);
+        // For each room, join the room as a member
+        roomsTransfer.forEach((roomId) => {
+          RoomMembersCollection.addMember(roomId, adminId);
+        });
+
+        // Transfer Boards
+        const boardsTransfer = await BoardsCollection.transferUsersBoards(userIdToDelete, adminId);
+
+        // Transfer Apps
+        const appsTransfer = await AppsCollection.transferUsersApps(userIdToDelete, adminId);
+
+        // Transfer Assets
+        const assetsTransfer = await AssetsCollection.transferUsersAssets(userIdToDelete, adminId);
+
+        // Transfer Plugins
+        const pluginsTransfer = await PluginsCollection.transferUsersPlugins(userIdToDelete, adminId);
+
+        // Log the transfer information
+        this.userDeleteLog(`Rooms Transfer: ${roomsTransfer ? 'Success' : 'Failed'}`, userEmail);
+        this.userDeleteLog(`Boards Transfer: ${boardsTransfer ? 'Success' : 'Failed'}`, userEmail);
+        this.userDeleteLog(`Apps Transfer: ${appsTransfer ? 'Success' : 'Failed'}`, userEmail);
+        this.userDeleteLog(`Assets Transfer: ${assetsTransfer ? 'Success' : 'Failed'}`, userEmail);
+        this.userDeleteLog(`Plugins Transfer: ${pluginsTransfer ? 'Success' : 'Failed'}`, userEmail);
+
+        this.userDeleteLog('User Deletion Request Finished', userEmail);
+        res.status(200).send({ success: true });
       } else {
         this.userDeleteLog('User data will be permentately deleted', userEmail);
